windows 平台兼容api开发windows 平台兼容(编译三方库的windows版)

+++ framework/mem/shm_api.h
@@ -36,7 +36,7 @@ namespace ShareMemAPI
 	// Parameter: SMKey iKey
 	// Parameter: int vSize
 	//************************************
-	SMKey CreateShareMem(SMKey iKey, int vSize);
+	SMKey CreateShareMem(SMKey iKey, size_t vSize);
 	//************************************
 	// Method:    OpenShareMem
 	// FullName:  ShareMemAPI::OpenShareMem
@@ -46,7 +46,7 @@ namespace ShareMemAPI
 	// Parameter: SMKey iKey
 	// Parameter: int vSize
 	//************************************
-	SMKey OpenShareMem(SMKey iKey, int vSize);
+	SMKey OpenShareMem(SMKey iKey, size_t vSize);
 	//************************************
 	// Method:    AttachShareMem
 	// FullName:  ShareMemAPI::AttachShareMem
